ui: refactor; extract more job components
Duration, FractionCompleted, HighwaterTimestamp, Progress
components have been extracted from JobStatusCell

Release note: none
