+++ pkg/ui/src/views/jobs/jobStatusOptions.ts
@@ -1,3 +1,6 @@
+import {cockroach} from "src/js/protos";
+import Job = cockroach.server.serverpb.JobsResponse.IJob;
+
 export const statusOptions = [
   {value: "", label: "All"},
   {value: "pending", label: "Pending"},
@@ -7,3 +10,7 @@ export const statusOptions = [
   {value: "succeeded", label: "Succeeded"},
   {value: "failed", label: "Failed"},
 ];
+
+export function jobHasOneOfStatuses(job: Job, ...statuses: string[]) {
+  return statuses.indexOf(job.status) !== -1;
+}
